refactor: move newtypes into module
